Functioning Server! Two existing routes:
generateRandomOutfits, which accepts a gender for top bottom and shoes as query arguments and returns a list of 30 outfits.

generateRandomOutfitsByBrand, which accepts a gener for top bottom and shoes, as well as a brand as query arguments, It returns 15 random fits with items from that brand, filling in with random clothing if there aren't 15 documents with that brand.